third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/svg/text/scripted/lengthadjust-expected.txt

This is a testharness.js-based test.
FAIL Tests dynamic updates of the "lengthAdjust" content attribute assert_greater_than: expected a number greater than 86 but got 86
FAIL Tests dynamic updates of the "lengthAdjust" IDL attribute assert_greater_than: expected a number greater than 86 but got 86
Harness: the test ran to completion.
